Backend/requirements.txt.txt
Flask==2.2.5
Flask-SQLAlchemy==3.0.5
Flask-Mail==0.9.1
Flask-Cors==3.0.10
numpy==1.24.3
pandas==1.5.3
tensorflow==2.11.0
python-dotenv==1.0.0
Werkzeug==2.2.3
